Protected gps logs
Create new table prot that contains filename, time, value.
When a file is protected, gps logs that are applicable to it are
copied into prot. When unprotected, they are deleted from prot.

Also updated the reaper to delete gps logs older than the last file
in the filesystem. I'm not bothering to make this work with
standalone non-RTC mode, since if it is possible to make a gps log
then you know the real time.