Mutex protected FS readwrite control and database access.
Add a counter to determine the number of "users" there are for the
writable filesystem and the database.

Each item that requires use of writable filesystem or database must
increment the counter when they obtain access, and decrement the
counter when they release access.

The counter is mutex protected.

When the filesystem counter reaches 0, it is remounted RO.
When the filesystem counter increments from 0, it is remounted RW.
When the database counter reaches 0, it is closed.
When the database counter increments from 0, it is opened.

The database is a user of the writable filesystem, so when anything
requests database access, by extension, it also requests writable
filesystem access.

This functionality is managed by;
pthread_mutex_t fs_mutex, db_mutex;
int fs_users, db_users,

Functions;
isFilesystemRO()
remountFS(int)
getWritableFS()
releaseWritableFS()
getGPSDB()
releaseGPSDB()